Refactor API calls and remove unused teaApi and store files; update RegisterForm to simplify state management
