--- src/app/store.js
@@ -1,1 +1,0 @@
-// This file has been removed as part of simplifying state management
